dmistData {viperData}	R Documentation

DMIST data

Description

DMIST performance data manually transcribed from Pisano2005_NEJM_v353p1773 and its supplement.

Usage


dmistData

Format

An object of class list of length 6.

Details

The performance data is organized in a list with the following elements.

nObs.FFDM [num] The number of FFDM observations (from Supplemental Table 2)

nObs.SFM [num] The number of SFM observations (from Supplemental Table 2)

nObs.total [num] the total number of observations (from Supplemental Table 2)

caseDist [list] This list contains two items (from Supplemental Table 2)

desc [char] A description of the data.

caseDist.df [data frame] The distribution of BIRADS scores from DMIST.
Row 1 is the distribution of FFDM BIRADS scores among DMIST cancers.
Row 2 is the distribution of SFM BIRADS scores among DMIST cancers.
Row 3 is the distribution of FFDM BIRADS scores among all DMIST cases.
Row 4 is the distribution of SFM BIRADS scores among all DMIST cases.

caseDist.denseBreasts [data frame] Caption to Figure 1 and row "T stage" of Table 2. One row of cancer=n1 detected by FFDM and SFM, cancer=n1 detected by SFM but not FFDM, cancer=n1 detected by FFDM but not SFM, cancer=n1 not detected by SFM or FFDM, non-cancer=n2, and total.

performance [list]

desc [char] A description of the data taken from Figure 1 and the text of Pisano2005_NEJM_v353p1773.pdf. SE of AUC is taken to be 1/4 of the confidence interval specified. AUC determined using seven-point malignancy scale. TPF,FPF are from Table 2 of Pisano2005_NEJM_v353p1773-Suppl.pdf. TPF,FPF are determined using BIRADS. Cancer determined at 455 days for AUC, TPF, and FPF.

FFDM [list] A list of the diagnostic performance statistics: AUC, AUCse, CIbot, CItop, TPF, TPFse, TNF, TNFse

SFM [list] A list of the diagnostic performance statistics: AUC, AUCse, CIbot, CItop, TPF, TPFse, TNF, TNFse

FFDMminusSFM [list] A list of the diagnostic performance statistics: AUC, AUCse, CIbot, CItop, p, TPF, TPFse, TPFCIbot, TPFCItop, TPFp, TNF, TNFse, TNFCIbot, TNFCItop, TNFp

[Package viperData version 1.0.0 Index]