dmistRSsummary {viperData}	R Documentation

dmistRSsummary

Description

Performance summaries from the following DMIST reader studies:

Hendrick2008_Radiology_v247p38, 'Accuracy of soft-copy digital mammography versus that of screen-film mammography according to digital manufacturer: ACRIN DMIST retrospective multireader study.' There were four machines studied, labeled below as hendrick2008.GE, hendrick2008.Fuji, hendrick2008.Fischer, hendrick2008.Hologic. The summary based on the aggregate of all this data is hendrick2008.

Nishikawa2009_Radiology_v251p41, 'Comparison of Soft-Copy and Hard-Copy Reading for Full-Field Digital Mammography.' There were four machines studied, labeled below as nishikawa2009.GE, nishikawa2009.Fuji, nishikawa2009.Fischer, nishikawa2009.Hologic. The summary based on the aggregate of all this data is nishikawa2009.

An unpublished study evaluating the effect of breast density on diagnostic accuracy of FFDM and SFM. The study included 320 women (109 cancers) stratified by original SFM breast density determination, 8 readers reading FFDM, and 8 readers reading SFM. There were four density levels studied, corresponding to the four BIRADS density levels, labeled as breastDensity.A, breastDensity.B, breastDensity.C, breastDensity.D. The summary based on the aggregate of all this data is breastDensity.

An unpublished study evaluating performance in a reader study at 50 30 readers assigned to read a total of 300 cases as FFDM and SFM. There was one data set, labeled as prevalence50.

Usage


dmistRSsummary

Format

An object of class list of length 16.

[Package viperData version 1.0.0 Index]