doViperStudyDesign {viperData}	R Documentation

doViperStudyDesign

Description

Extract basic study design info

Usage


doViperStudyDesign(rawData)

Arguments

rawData
One of the VIPER sub-study data frames.

Value

A list object that contains several ways to summarize the distribution of cases in the study accoring to case type.

nObs [num] The number of observations in the sub-study.

readers [char] The reader IDs in the sub-study.

nR [num] The number of readers in the sub-study.

nC [data frame] The number of cases in the sub-study by case type. One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

perReader [list] This list contains two data frames: FFDM and SFM. Each of these data frames includes the number of cases evaluated by each reader by case type. 20 readers == 20 Rows of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

perReader.summary [list] This list contains two data frames: FFDM and SFM. Each of these data frames includes the total number of cases evaluated by all readers by case type. One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

Examples


df <- split(viperData::viperObs455, viperData::viperObs455$desc)
df.studyDesign <- doViperStudyDesign(df$screeningLowP)

[Package viperData version 1.0.0 Index]