viperObs {viperData}	R Documentation

viperObs

Description

The VIPER data sets were created by aggregating all the raw reader data files from the FDA Validation of Imaging Premarket Evaluation and Regulation (VIPER): viperObs365 and viperObs455.

Details

The images used in this study were selected from the DMIST images [Pisano2005_NEJM_v353p1773]. The VIPER study concept 2011. Data collection September 2013 to August 2015. Please refer to journal paper for details.

viperObs is a data frame with 12 variables:

caseType455 [str] Cancer type based on 455 days result: birads123sfm, birads0sfm, birads0ffdm, birads123ffdm, cancer

caseType365 [str] Cancer type based on 365 days result: birads123sfm, birads0sfm, birads0ffdm, birads123ffdm, cancer

cancerStatus455 [num] Cancer status based on 455 days result. 1: Cancer; 2: Not cancer

cancerStatus365 [num] Cancer status based on 365 days result. 1: Cancer; 2: Not cancer

caseID [num] case ID

readerID [str] reader ID: 44 readers

modalityID [str] modality ID: FFDM, SFM

recall [num] binary recall decision. 1: recall; 2: not recall

abnormalityType abnormality type of case

score [num] dignostic result
0-100: not recall, lower value means higher not recall confidence
101-200: recall, higer value means higher recall confidence

caseGroup [str] group of case: cgrp1, cgrp2, cgrp3, cgrp4

subStudyLabel [str] study group: screeningLowP,screeningMedP,screeningHighP, challengeMedP,challengeHighP

[Package viperData version 1.0.2 Index]