viperObservations {viperData}	R Documentation

viperObservations

Description

The VIPER data sets were created by aggregating all the raw reader data files from the FDA VIPER study: Validation of Imaging Premarket Evaluation and Regulation (VIPER). The data sets differ in the truth labels. See the documentation details.

Usage


viperObs365

viperObs455

Format

An object of class data.frame with 20382 rows and 35 columns.

Details

A data frame with 36 variables:

CaseReadOrder

CaseID

Modality

ReaderID

Study

Track

ReadSetLetter

Ctype

Compare.CaseIDs

Initial.Comparison

Final.Comparison

case_id

redcap_event_name

readerID

form_start_datetime

modality

study

viewbox

num_img

recall

miss_img

score_start_datetime

abnormality_type

recall_scale

score_end_datetime

Group

norecall_scale

set_letter

track

quality

UID

score

desc

read_order

Comment

RecallRegion

count

viperObs365: This dataset is identical to viperObs455 except for the truth labels. The truth labels for this dataset are based on cancer at 365 days.

viperObs455: This dataset is identical to viperObs365 except for the truth labels. The truth labels for this dataset are based on cancer at 455 days.

The images used in this study were selected from the DMIST images [Pisano2005_NEJM_v353p1773]. The VIPER study concept 2011. Data collection September 2013 to August 2015. Please refer to journal paper for details.

[Package viperData version 1.0.0 Index]