viperReaderQualifications {viperData}	R Documentation

VIPER reader qualification summary

Description

This data set contains the information from the VIPER reader recruitment qualification forms. The original data was modified to turn text in numeric fields to appropriate numbers. For example, "~100" was modified to 100, "100+" was modified to 100, "<50" was modified to 50, "200-250" was modified to 225, "don't know" was modified to "NA", and "very few" was modified to "NA".

Usage


viperReaderQualifications

Format

An object of class data.frame with 43 rows and 10 columns.

Details

A data frame with 10 variables:

readerID

certifiedABR

breastImagingFellowship

fellowshipYear

min50sfm

min50ffdm

howManySFMlast2years

howManyFFDMlast2years

howManyYearsPostResidencyExperienceInterpretingSFM

howManyYearsPostResidencyExperienceInterpretingFFDM

[Package viperData version 1.0.0 Index]