viperSummary {viperData}	R Documentation

viperSummaries

Description

The VIPER summary files contain summary statistics of the VIPER observations. The iMRMC R package is used to estimate the area under the ROC curve (AUC), the true positive fraction (sensitivity), and the true negative fraction (specificity). The iMRMC package also estimates variances, standard errors, and confidence intervals that account for the variability from readers and cases. You can refer to the R function doViperSummary.R in this package for the code that produces the summary files.

The iMRMC R package can be found at

https://github.com/DIDSR/iMRMC/releases,

https://github.com/DIDSR/iMRMC/tree/master/Rpackage, and

https://cran.r-project.org/web/packages/iMRMC/index.html.

Usage


viperSummary365

viperSummary455

Format

An object of class list of length 5.

Details

Each element of the list pertains to one of the VIPER sub-studies:

screeningLowP [list] Screening population, per reader prevalence ~ 10%

screeningMedP [list] Screening population, per reader prevalence ~ 30%

screeningHighP [list] Screening population, per reader prevalence ~ 50%

challengeMedP [list] Challenge population, per reader prevalence ~ 30%

challengeHighP [list] Challenge population, per reader prevalence ~ 50%

Each substudy list element is a list object of length 3.

desc [string] Sub-study label. This is the same as the sub-study list element name.

studyDesign [list] This list object contains several ways to summarize the distribution of cases in the study accoring to case type.

nObs [num] The number of observations in the sub-study.

readers [char] The reader IDs in the sub-study.

nR [num] The number of readers in the sub-study.

nC [data frame] The number of cases in the sub-study by case type. One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

perReader [list] This list contains two data frames: FFDM and SFM. Each of these data frames includes the number of cases evaluated by each reader by case type. 20 readers == 20 Rows of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

perReader.summary [list] This list contains two data frames: FFDM and SFM. Each of these data frames includes the total number of cases evaluated by all readers by case type. One row of birads0ffdm, birads0sfm, birads12ffdm, birads12sfm, cancer, noncancer.

iMRMC [list] This list object contains the MRMC analysis results for auc, tpf, and tnf. There are four groups of results: perReader, Ustat, MLEstat, ROC, and varDecomp The analysis results are produced by a command line version of iMRMC called by R: https://github.com/DIDSR/iMRMC/releases, https://github.com/DIDSR/iMRMC/tree/master/Rpackage, https://cran.r-project.org/web/packages/iMRMC/index.html

viperSummary365: This summary is based on viperObs365 the dataset based on cancer at 365 days.

viperSummary455: This summary is based on viperObs455 the dataset based on cancer at 455 days.

[Package viperData version 1.0.0 Index]